move textureutils to lib/utils
