+++ lib/util/textureutils.ts
